Moved shoot ray to physics module
