+++ engine/physics/public/physics_module.hpp
@@ -5,6 +5,7 @@
 #pragma once
 
 #include "Jolt/Jolt.h"
+#include "entt/entity/entity.hpp"
 #include <Jolt/Physics/Collision/CastResult.h>
 #include <Jolt/Physics/Collision/CollisionCollectorImpl.h>
 #include <Jolt/Physics/Collision/RayCast.h>
@@ -244,12 +245,23 @@ private:
     std::vector<glm::vec3> persistentLinePositions;
 };
 
+struct RayHitInfo
+{
+    entt::entity entity = entt::null; // Entity that was hit
+    glm::vec3 position = glm::vec3(0.0f, 0.0f, 0.0f); // Position where the ray hits; HitPoint = Start + mFraction * (End - Start)
+    float hitFraction = 0.0f; // Hit fraction of the ray/object [0, 1], HitPoint = Start + mFraction * (End - Start)
+    bool hasHit = false;
+};
+
 class PhysicsModule
 {
 public:
     PhysicsModule();
     ~PhysicsModule();
     void UpdatePhysicsEngine(float deltaTime);
+
+    NO_DISCARD RayHitInfo ShootRay(const glm::vec3& origin, const glm::vec3& direction, float distance) const;
+
     JPH::BodyInterface* bodyInterface = nullptr;
     DebugRendererSimpleImpl* debugRenderer = nullptr;
     JPH::PhysicsSystem* physicsSystem = nullptr;
